ultim commit del sprint 5

+++ app/Livewire/Medias/IndexMedias.php
@@ -9,28 +9,54 @@ use Livewire\Component;
 class IndexMedias extends Component
 {
     public $medias;
+
+    public $selectedRows = [];
+
     public function mount()
     {
         $this->medias = Media::all();
     }
+
+    public function cridaSave()
+    {
+        return redirect()->to('/medias/create');
+    }
+
+    public function cridaUpdate($id)
+    {
+        return redirect()->route('medias.update', ['id' => $id]);
+    }
+
+
     public function render()
     {
         return view('livewire.medias.index');
     }
 
-    public function delete($id)
+    public function delete()
     {
-        try {
-            $media = Media::findOrFail($id);
-            $media->delete();
-            //filtrem la media esborrada de la col·lecció de medias que
-            //passem a la vista
-            $this->medias = $this->medias->filter(fn($cat) => $cat->id
-                !== $id);
-            session()->flash('message', 'Media was successfully deleted.');
-        } catch (ModelNotFoundException $th) {
-            session()->flash('message-danger', 'Can not find the media to be deleted.');
+        foreach ($this->selectedRows as $id) {
+            $media = Media::find($id);
+            if ($media) {
+                $media->delete();
+                session()->flash('message', 'Media was successfully deleted.');
+            }
+        }
+        $this->selectedRows = [];
+        $this->mount();
+    }
+
+    public function selectAll()
+    {
+        if (count($this->selectedRows) < count($this->medias)) {
+            $this->selectedRows = $this->medias->pluck('id')->map(function ($id) {
+                return (string) $id;
+            })->toArray();
+        } else {
+            $this->selectedRows = [];
         }
     }
+
+
 
 }
